move channel support to Experimental namespace
