+++ examples/remotechannels/Client/Program.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using Messages;
 using Proto;
-using Proto.Channels;
+using Proto.Channels.Experimental;
 using Proto.Remote.GrpcNet;
 using static System.Threading.Channels.Channel;
 using static Proto.Remote.GrpcNet.GrpcNetRemoteConfig;
@@ -9,9 +9,10 @@ using static Proto.Remote.GrpcNet.GrpcNetRemoteConfig;
 var system = await ActorSystem.StartNew(Remote.Config(BindToLocalhost()));
 var publisher = PID.FromAddress("127.0.0.1:8000", "publisher");
 var channel = CreateUnbounded<MyMessage>();
-_ = ChannelSubscriber.StartNew(system.Root, publisher, channel);
+_ = channel.SubscribeToPid(system.Root, publisher);
 
 Console.WriteLine("Waiting for messages");
+
 await foreach (var msg in channel.Reader.ReadAllAsync())
 {
     Console.WriteLine($"Got message {msg.Value}");
