+++ examples/remotechannels/Server/Program.cs
@@ -2,14 +2,15 @@
 using System.Threading.Tasks;
 using Messages;
 using Proto;
-using Proto.Channels;
+using Proto.Channels.Experimental;
 using Proto.Remote.GrpcNet;
 using static Proto.Remote.GrpcNet.GrpcNetRemoteConfig;
 using static System.Threading.Channels.Channel;
 
 var system = await ActorSystem.StartNew(Remote.Config(BindToLocalhost(8000)));
 var channel = CreateUnbounded<MyMessage>();
-_ = ChannelPublisher.StartNew(system.Root, channel, "publisher");
+var publisher = system.Root.SpawnPublisherActor<MyMessage>("publisher");
+_ = channel.PublishToPid(system.Root, publisher);
 
 //produce messages
 for (var i = 0; i < 30; i++)
