+++ src/Proto.Actor/Channels/Experimental/ChannelSubscriberActor.cs
@@ -7,19 +7,8 @@ using System.Threading.Channels;
 using System.Threading.Tasks;
 using JetBrains.Annotations;
 
-namespace Proto.Channels
+namespace Proto.Channels.Experimental
 {
-    [PublicAPI]
-    public static class ChannelSubscriber
-    {
-        public static PID StartNew<T>(IRootContext context, PID publisher, Channel<T> channel)
-        {
-            var props = Props.FromProducer(() => new ChannelSubscriberActor<T>(publisher, channel));
-            var pid = context.Spawn(props);
-            return pid;
-        }
-    }
-    
     [PublicAPI]
     public class ChannelSubscriberActor<T> : IActor
     {
